ops/warehouse/requirements.txt
psycopg2-binary>=2.9
pdfminer.six>=20231228
